Uniform naming in repository classes
