+++ src/repositories/walletRepository.ts
@@ -11,20 +11,25 @@ export type CommitmentTxRecord = {
 };
 
 export interface WalletRepository extends AsyncDisposable {
+    /**
+     * Clear all data from storage.
+     */
+    clear(): Promise<void>;
+
     // VTXO management
     getVtxos(address: string): Promise<ExtendedVirtualCoin[]>;
     saveVtxos(address: string, vtxos: ExtendedVirtualCoin[]): Promise<void>;
-    clearVtxos(address: string): Promise<void>;
+    deleteVtxos(address: string): Promise<void>;
 
     // UTXO management
     getUtxos(address: string): Promise<ExtendedCoin[]>;
     saveUtxos(address: string, utxos: ExtendedCoin[]): Promise<void>;
-    clearUtxos(address: string): Promise<void>;
+    deleteUtxos(address: string): Promise<void>;
 
     // Transaction history
     getTransactionHistory(address: string): Promise<ArkTransaction[]>;
     saveTransactions(address: string, txs: ArkTransaction[]): Promise<void>;
-    clearTransactions(address: string): Promise<void>;
+    deleteTransactions(address: string): Promise<void>;
 
     // Wallet state
     getWalletState(): Promise<WalletState | null>;
